Just grab BOBBuilder from git
git-svn-id: svn+ssh://bbfs-01.calit2.net/grw/Gordon/svn/trunk/Gadgets/Tools/pyinstall@22403 2cf53259-273a-af40-bda6-69d31023b8ec

+++ gitinstall.sh
@@ -2,8 +2,8 @@
 
 # Envisioned workload:
 #  1.  Create a new development direcotry: 'mkdir Gadgetron'
-#  2.  git clone git@146.148.85.41:swanson/test-pyinstall.git
-#  3.  execute 'pyinstall/gitinstall.sh'
+#  2.  git clone git@github.com:NVSL/gadgetron-setup.git
+#  3.  execute 'gadgetron-setup/gitinstall.sh'
 #  4.  Update your PATH (this script prints out the command you should execute)
 
 #  You should have all the gadgetron tools in 'Gadgetron/' ready for your hacking/committing etc.  All dependencies should be properly tracked and installed a private virtualenv.
@@ -15,26 +15,31 @@ if [ ".$GADGETRON_ROOT" = "." ]; then
     GADGETRON_ROOT=$PWD
 fi
 
-if [ ".$PYTHON_ENV" = "." ]; then
-    PYTHON_ENV=$GADGETRON_ROOT/python_venv
+if [ ".$USE_VENV" != "." ]; then
+    
+    if [ ".$PYTHON_ENV" = "." ]; then
+	PYTHON_ENV=$GADGETRON_ROOT/python_venv
+    fi
+
+
+    if ! [ -e $PYTHON_ENV/bin/python ]; then
+	rm -rf $PYTHON_ENV
+	virtualenv $PYTHON_ENV
+    fi
+
+    PYTHON_ENV=$(cd $GADGETRON_ROOT/python_venv; pwd -P) # get the canonical path
+
+    PATH=$PYTHON_ENV/bin:$PATH
 fi
 
-
-if ! [ -e $PYTHON_ENV/bin/python ]; then
-    rm -rf $PYTHON_ENV
-    virtualenv $PYTHON_ENV
-fi
-
-PYTHON_ENV=$(cd $GADGETRON_ROOT/python_venv; pwd -P) # get the canonical path
-
-PATH=$PYTHON_ENV/bin:$PATH
-
-packages="test-Swoop test-Koala2"
+packages="BOBBuilder" #test-Swoop test-Koala2"
 
 for p in $packages; do
-    git clone ssh://git@146.148.85.41/swanson/${p}.git
-    (cd $p; python setup.py develop)
+    git clone git@github.com:NVSL/$p.git
+    (cd $p; make)
 done
 
-echo PATH=$PYTHON_ENV/bin:\$PATH
+if [ ".$USE_VENV" != "." ]; then
+    echo PATH=$PYTHON_ENV/bin:\$PATH
+fi
